Yet another rename!  This time to the approved (and final) name!  So now at 1.0.0rc4.
