+++ tests/test_actions.py
@@ -22,9 +22,9 @@ import mock
 import json
 import pytest
 
-from adobe_umapi_client import UMAPI, Action
-from adobe_umapi_client.error import UMAPIError, UMAPIRetryError, UMAPIRequestError, ActionFormatError
-from adobe_umapi_client.auth import Auth
+from umapi_client import UMAPI, Action
+from umapi_client.error import UMAPIError, UMAPIRetryError, UMAPIRequestError, ActionFormatError
+from umapi_client.auth import Auth
 
 
 # This method will be used by the mock to replace requests.get / requests.post
@@ -47,7 +47,7 @@ def mocked_requests_call(target, **kwargs):
         return MockResponse(404, {})
 
 
-@mock.patch('adobe_umapi_client.api.requests.get', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.get', side_effect=mocked_requests_call)
 def test_list_users_success(_):
     """Test Users List - SUCCESS"""
     auth = mock.create_autospec(Auth)
@@ -55,7 +55,7 @@ def test_list_users_success(_):
     assert api.users(None) == {"result": "success"}
 
 
-@mock.patch('adobe_umapi_client.api.requests.get', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.get', side_effect=mocked_requests_call)
 def test_list_users_error(_):
     """Test Users List - ERROR"""
     auth = mock.create_autospec(Auth)
@@ -63,7 +63,7 @@ def test_list_users_error(_):
     pytest.raises(UMAPIRequestError, api.users, None)
 
 
-@mock.patch('adobe_umapi_client.api.requests.get', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.get', side_effect=mocked_requests_call)
 def test_list_users_failure(_):
     """Test Users List - FAILURE"""
     auth = mock.create_autospec(Auth)
@@ -73,7 +73,7 @@ def test_list_users_failure(_):
     pytest.raises(UMAPIRetryError, api.users, None)
 
 
-@mock.patch('adobe_umapi_client.api.requests.get', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.get', side_effect=mocked_requests_call)
 def test_list_groups_success(_):
     """Test Groups List - SUCCESS"""
     auth = mock.create_autospec(Auth)
@@ -81,7 +81,7 @@ def test_list_groups_success(_):
     assert api.groups(None) == {"result": "success"}
 
 
-@mock.patch('adobe_umapi_client.api.requests.post', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.post', side_effect=mocked_requests_call)
 def test_user_create_success(_):
     """Test User Creation - SUCCESS"""
     auth = mock.create_autospec(Auth)
@@ -94,7 +94,7 @@ def test_user_create_success(_):
     assert api.action(None, action) == {"result": "success"}
 
 
-@mock.patch('adobe_umapi_client.api.requests.post', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.post', side_effect=mocked_requests_call)
 def test_user_create_error(_):
     """Test User Creation - ERROR"""
     auth = mock.create_autospec(Auth)
@@ -106,7 +106,7 @@ def test_user_create_error(_):
     pytest.raises(UMAPIRequestError, api.action, None, action)
 
 
-@mock.patch('adobe_umapi_client.api.requests.post', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.post', side_effect=mocked_requests_call)
 def test_user_create_failure(_):
     """Test User Creation - FAILURE"""
     auth = mock.create_autospec(Auth)
@@ -121,7 +121,7 @@ def test_user_create_failure(_):
     pytest.raises(UMAPIRetryError, api.action, None, action)
 
 
-@mock.patch('adobe_umapi_client.api.requests.post', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.post', side_effect=mocked_requests_call)
 def test_product_add(_):
     """Test Product Add - SUCCESS"""
     auth = mock.create_autospec(Auth)
@@ -134,7 +134,7 @@ def test_product_add(_):
     assert api.action(None, action) == {"result": "success"}
 
 
-@mock.patch('adobe_umapi_client.api.requests.post', side_effect=mocked_requests_call)
+@mock.patch('umapi_client.api.requests.post', side_effect=mocked_requests_call)
 def test_action_format_error(_):
     """Test Action Format Error"""
     auth = mock.create_autospec(Auth)
